typecast spaces_remaining var as int in case identical comparison is not triggering because of it being a string [touch:4705]

+++ core/db_classes/EE_Event.class.php
@@ -554,7 +554,7 @@ class EE_Event extends EE_CPT_Base{
 			if ( $datetime->spaces_remaining() === INF ) {
 				return;
 			} else {
-				$spaces_remaining = max( $datetime->spaces_remaining(), $spaces_remaining );
+				$spaces_remaining = (int)max( $datetime->spaces_remaining(), $spaces_remaining );
 			}			
 		}
 		if ( $spaces_remaining === 0 ) {
